db-load/requirements.txt
psycopg2-binary
pandas
sqlalchemy
python-dotenv